requirements.txt
fastapi
uvicorn[standard]
transformers
python-docx
PyPDF2
python-pptx
pandas
torch
tqdm
python-multipart
huggingface_hub[hf_xet]
streamlit
requests